Move SearchAction declaration to searchaction.h
