+++ src/ui/cli/search.h
@@ -9,28 +9,9 @@
 #ifndef __pwsafe_xcode6__search__
 #define __pwsafe_xcode6__search__
 
-#include <string>
-#include "../../os/UUID.h"
-#include "../../core/PWScore.h"
-#include "./argutils.h"
+class PWScore;
+class UserArgs;
 
-
-struct SearchAction
-{
-  SearchAction() {}
-  virtual void operator()(const pws_os::CUUID &uuid, const CItemData &data) = 0;
-  virtual int Execute() = 0;
-  
-  SearchAction& operator=(const SearchAction&) = delete;
-  SearchAction& operator=(const SearchAction&&) = delete;
-  SearchAction( const SearchAction& ) = delete;
-  SearchAction( const SearchAction&& ) = delete;
-  
-};
-
-void SearchForEntries(PWScore &core, const std::wstring &searchText, bool ignoreCase,
-                      const Restriction &restrictToEntries, const CItemData::FieldBits &fieldsToSearch,
-                      SearchAction &cb);
 int Search(PWScore &core, const UserArgs &ua);
 int SaveAfterSearch(PWScore &core, const UserArgs &ua);
